+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## [3.0.3] - 2022/09/14
+## [4.0.0] - 2022/12/14
+* refactor buttons
+* add style.elevation
+* add button aspectRatio
+
+## [3.0.3] - 2022/12/01
 * refactor
 * Add custom Icon